Create test module and remove src folder
Modifies the test directory to be a Bazel module. This allows validating that the rules_browsers module works when depended on by another module/project. Making this work required some small changes to pathing logic.

Bazel modules typically provide rules in top level folders and not in a single src. This removes the src folder and moves its contents to the root. This makes imports easier for dependents.