Remove internal dependency on browsers.
This is a major refactor that removes the dependency of `rules_browsers` on the browsers it provides. Instead, dependents of `rules_browsers` would use the extension to download browsers as needed and wrap them with `browser_group` if they want to.

This changes the public API surface notably, but it also enables consumers to define which browsers to download and which to skip. All macros that require browsers now accept the browser as an attribute rather than hard-coding an internal reference.

Overall this will require consumers to make changes but it's much more versatile.

One more note: This is required to submit to the BCR, which had complained about having each use of the module download all browsers. That's a sensible complaint I'd say, but it requires this major change.